+++ packages/components/datepicker/CHANGELOG.md
@@ -3,6 +3,190 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0-alpha.70](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.69...@contentful/forma-36-react-datepicker@0.4.0-alpha.70) (2021-08-04)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.69](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.68...@contentful/forma-36-react-datepicker@0.4.0-alpha.69) (2021-08-03)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.68](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.67...@contentful/forma-36-react-datepicker@0.4.0-alpha.68) (2021-08-03)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.67](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.66...@contentful/forma-36-react-datepicker@0.4.0-alpha.67) (2021-07-29)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.66](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.65...@contentful/forma-36-react-datepicker@0.4.0-alpha.66) (2021-07-28)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.65](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.64...@contentful/forma-36-react-datepicker@0.4.0-alpha.65) (2021-07-28)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.64](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.63...@contentful/forma-36-react-datepicker@0.4.0-alpha.64) (2021-07-23)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.63](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.62...@contentful/forma-36-react-datepicker@0.4.0-alpha.63) (2021-07-22)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.62](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.61...@contentful/forma-36-react-datepicker@0.4.0-alpha.62) (2021-07-22)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.61](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.60...@contentful/forma-36-react-datepicker@0.4.0-alpha.61) (2021-07-22)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.60](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.59...@contentful/forma-36-react-datepicker@0.4.0-alpha.60) (2021-07-19)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.59](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.58...@contentful/forma-36-react-datepicker@0.4.0-alpha.59) (2021-07-19)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.58](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.57...@contentful/forma-36-react-datepicker@0.4.0-alpha.58) (2021-07-19)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.57](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.56...@contentful/forma-36-react-datepicker@0.4.0-alpha.57) (2021-07-14)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.56](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.55...@contentful/forma-36-react-datepicker@0.4.0-alpha.56) (2021-07-13)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.55](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.54...@contentful/forma-36-react-datepicker@0.4.0-alpha.55) (2021-07-06)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.54](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.53...@contentful/forma-36-react-datepicker@0.4.0-alpha.54) (2021-07-02)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.53](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.52...@contentful/forma-36-react-datepicker@0.4.0-alpha.53) (2021-07-01)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.52](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.51...@contentful/forma-36-react-datepicker@0.4.0-alpha.52) (2021-06-29)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.51](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.50...@contentful/forma-36-react-datepicker@0.4.0-alpha.51) (2021-06-28)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.50](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.49...@contentful/forma-36-react-datepicker@0.4.0-alpha.50) (2021-06-28)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.49](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.48...@contentful/forma-36-react-datepicker@0.4.0-alpha.49) (2021-06-17)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
+# [0.4.0-alpha.48](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.47...@contentful/forma-36-react-datepicker@0.4.0-alpha.48) (2021-06-07)
+
+**Note:** Version bump only for package @contentful/forma-36-react-datepicker
+
+
+
+
+
 # [0.4.0-alpha.47](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-datepicker@0.4.0-alpha.46...@contentful/forma-36-react-datepicker@0.4.0-alpha.47) (2021-05-26)
 
 **Note:** Version bump only for package @contentful/forma-36-react-datepicker