+++ packages/forma-36-fcss/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.3](https://github.com/contentful/forma-36/compare/@contentful/forma-36-fcss@0.3.2...@contentful/forma-36-fcss@0.3.3) (2021-07-22)
+
+**Note:** Version bump only for package @contentful/forma-36-fcss
+
+
+
+
+
 ## [0.3.2](https://github.com/contentful/forma-36/compare/@contentful/forma-36-fcss@0.3.1...@contentful/forma-36-fcss@0.3.2) (2021-05-07)
 
 **Note:** Version bump only for package @contentful/forma-36-fcss